Relaxed command line argument requirements from `hummingbot_quickstart.py`... to allow better integration with current Docker installation instructions.
