+++ bin/hummingbot_quickstart.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import path_util        # noqa: F401
+import argparse
 import asyncio
 import logging
 from typing import (
@@ -8,6 +9,7 @@ from typing import (
     List,
 )
 import os
+
 from hummingbot import (
     check_dev_mode,
     init_logging,
@@ -21,8 +23,8 @@ from hummingbot.client.config.config_helpers import (
     update_strategy_config_map_from_file,
     all_configs_complete,
 )
+from hummingbot.client.ui import login_prompt
 from hummingbot.client.ui.stdout_redirection import patch_stdout
-from hummingbot.client.ui.parser import ThrowingArgumentParser
 from hummingbot.client.settings import STRATEGIES
 from hummingbot.core.utils.async_utils import safe_gather
 from hummingbot.core.management.console import start_management_console
@@ -35,17 +37,17 @@ from hummingbot.core.utils.exchange_rate_conversion import ExchangeRateConversio
 from hummingbot.client.config.security import Security
 
 
-class CmdlineParser(ThrowingArgumentParser):
+class CmdlineParser(argparse.ArgumentParser):
     def __init__(self):
         super().__init__()
         self.add_argument("--strategy", "-s",
                           type=str,
                           choices=STRATEGIES,
-                          required=True,
+                          required=False,
                           help="Choose the strategy you would like to run.")
         self.add_argument("--config-file-name", "-f",
                           type=str,
-                          required=True,
+                          required=False,
                           help="Specify a file in `conf/` to load as the strategy config file.")
         self.add_argument("--wallet", "-w",
                           type=str,
@@ -57,15 +59,16 @@ class CmdlineParser(ThrowingArgumentParser):
                           help="Specify the password to unlock your encrypted files and wallets.")
 
 
-async def quick_start():
-    args = CmdlineParser().parse_args()
+async def quick_start(args):
     strategy = args.strategy
     config_file_name = args.config_file_name
     wallet = args.wallet
     password = args.config_password
-    if not Security.login(password):
+
+    if password is not None and not Security.login(password):
         logging.getLogger().error(f"Invalid password.")
         return
+
     try:
         await Security.wait_til_decryption_done()
         await create_yml_files()
@@ -76,9 +79,11 @@ async def quick_start():
         await ExchangeRateConversion.get_instance().wait_till_ready()
         hb = HummingbotApplication.main_application()
         # Todo: validate strategy and config_file_name before assinging
-        hb.strategy_name = strategy
-        hb.strategy_file_name = config_file_name
-        update_strategy_config_map_from_file(os.path.join(CONF_FILE_PATH, config_file_name))
+
+        if config_file_name is not None and strategy is not None:
+            hb.strategy_name = strategy
+            hb.strategy_file_name = config_file_name
+            update_strategy_config_map_from_file(os.path.join(CONF_FILE_PATH, config_file_name))
 
         # To ensure quickstart runs with the default value of False for kill_switch_enabled if not present
         if not global_config_map.get("kill_switch_enabled"):
@@ -87,12 +92,14 @@ async def quick_start():
         if wallet and password:
             global_config_map.get("ethereum_wallet").value = wallet
 
-        if not all_configs_complete(hb.strategy_name):
-            await hb.notify_missing_configs()
-            # config_map = load_required_configs()
-            # empty_configs = [key for key, config in config_map.items() if config.value is None and config.required]
-            # empty_config_description: str = "\n- ".join([""] + empty_configs)
-            # raise ValueError(f"Missing configuration values: {empty_config_description}\n")
+        if hb.strategy_name and hb.strategy_file_name:
+            if not all_configs_complete(hb.strategy_name):
+                await hb.notify_missing_configs()
+                # config_map = load_required_configs()
+                # empty_configs = [key for key, config in config_map.items()
+                #                  if config.value is None and config.required]
+                # empty_config_description: str = "\n- ".join([""] + empty_configs)
+                # raise ValueError(f"Missing configuration values: {empty_config_description}\n")
 
         with patch_stdout(log_field=hb.app.log_field):
             dev_mode = check_dev_mode()
@@ -102,10 +109,11 @@ async def quick_start():
             log_level = global_config_map.get("log_level").value
             init_logging("hummingbot_logs.yml",
                          override_log_level=log_level,
-                         dev_mode=dev_mode,
-                         strategy_file_path=config_file_name)
-            await write_config_to_yml(hb.strategy_name, hb.strategy_file_name)
-            hb.start(log_level)
+                         dev_mode=dev_mode)
+
+            if hb.strategy_file_name is not None and hb.strategy_name is not None:
+                await write_config_to_yml(hb.strategy_name, hb.strategy_file_name)
+                hb.start(log_level)
 
             tasks: List[Coroutine] = [hb.run()]
             if global_config_map.get("debug_console").value:
@@ -122,5 +130,15 @@ async def quick_start():
             raise e
 
 
+def main():
+    # If no password is given from the command line, prompt for one.
+    args = CmdlineParser().parse_args()
+    if args.config_password is None:
+        if not login_prompt():
+            return
+
+    asyncio.get_event_loop().run_until_complete(quick_start(args))
+
+
 if __name__ == "__main__":
-    asyncio.get_event_loop().run_until_complete(quick_start())
+    main()
